Make up/down/remove buttons behave sane on …AddParameterPage
I realised these are vital information to make the buttons at
the bottom of the template dialog behave sane. It's still
possible to focus this page, even if it doesn't have a visible
item in any of the old/new sidebars. This is when these flags
are used to decide if the up/down/remove buttons should be
enabled.

Bug: T291151
Change-Id: I6ab709b856d110bfb37daa1592c0b6a99714aa25