Improve messages about duplicate template parameters
There are some combinations that can result in a message that is
more confusing than helpful. For example, when a parameter is
already in a template, but via an alias, that alias is potentially
shown in the sidebar (only when the parameter doesn't have a
label). When the user enters the primary name of this parameter,
the message talks about a "duplicate", but the user can't
understand what this duplicate is.

Just make all messages repeat the parameter's label as shown in
the sidebar. This is often redundant, but sometimes crucial
information.

This patch also improves the qqq.json documentation. All messages
get 2 parameters, even if some of them use only 1.

Bug: T285869
Bug: T291059
Change-Id: I826e341cebd8f89dba63d64de67baba1ec8f2001