Remove unused media info messages
The "visualeditor-dialog-media-info-…" messages are part of a larger
info panel that appears after searching for an image. This was
introduced via I7b4d019, a patch with 52 patchsets, developed between
September 2014 and January 2015.

The "credit" message appeared first in patchset 2, see
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/VisualEditor/+/161342/2/modules/ve-mw/i18n/en.json

The "dateformat" message appeared first in patchset 12, see
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/VisualEditor/+/161342/11..12/modules/ve-mw/i18n/en.json
This message appears to be special in so far that it was never used
to begin with.

The other two messages appeared first in patchset 21, see
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/VisualEditor/+/161342/20..21/modules/ve-mw/i18n/en.json

3 of the 4 messages are used at this point. The message keys are
constructed in code via `"visualeditor-dialog-media-info-" + key`.
This piece of code was also introduced in patchset 21, see
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/VisualEditor/+/161342/20..21/modules/ve-mw/ui/dialogs/ve.ui.MWMediaDialog.js#629

This was entirely rewritten in patchset 41, see
https://gerrit.wikimedia.org/r/c/mediawiki/extensions/VisualEditor/+/161342/40..41/modules/ve-mw/ui/dialogs/ve.ui.MWMediaDialog.js#b670
Since then the messages are unused. The info panel still exists and
still shows all information, but the 3 fields "ImageDescription",
"LicenseShortName", and "Credit" just show an icon and the field's
value, but no label (any more).

Change-Id: Ief35a85d8c5cbf49b662adbfd068dbecd8e4f154