Remove unused media position messages
All 4 messages are introduced via I15fbd1e in January 2014.

The "none" message is unused since Iddd78c9 from February 2014. It
was replaced with a checkbox with a different label.

The other 3 messages are unused since I7057997 from February 2015.
A more generic widget is used since then. It uses messages that are
named "visualeditor-align-widget-center" and so on.

Change-Id: I7ad4772cb4fdec7e478cee23dccd5ae600834caf