Remove CSS applied to .ve-ui-mwTransclusionDialog
This class was (accidentally?) removed in 2014 so the styles had
no effect since then. I decided on removing the code that was
unused for 8years. - Alternativly the class could be re-added but
I'm not sure if it makes sense.

See I4b2ba31bed5c4f80940623702d635cacd19e0a66 where this got lost.

Change-Id: Ifcbfc4273e41c08431c6f5b0ca2f2b6d25c34edd